examples/requirements.txt
tiktoken
imageio
transformers
einx
dm-haiku
flax
torch
torchvision
opencv-python